requirements.txt
absl-py==0.12.0
asgiref==3.3.4
astor==0.8.1
astunparse==1.6.3
atari-py==0.2.9
attrs==21.2.0
autokeras==1.0.14
azure-core==1.15.0
azure-eventhub==5.1.0
azure-storage-blob==12.6.0
backcall==0.2.0
blessings==1.7
boltons==21.0.0
boto==2.49.0
boto3==1.17.88
botocore==1.20.88
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.3
cryptography==3.4.7
curtsies==0.3.5
cwcwidth==0.1.4
cycler==0.10.0
dask==2.6.0
decorator==4.4.2
Deprecated==1.2.12
dill==0.3.3
Django==3.2
django-cors-headers==3.7.0
django-http-exceptions==1.4.0
djangorestframework==3.12.4
fire==0.4.0
Flask==2.0.1
flatbuffers==1.12
fsspec==2021.5.0
gast==0.3.3
gitdb==4.0.7
GitPython==3.1.17
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
GPUtil==1.4.0
greenlet==1.1.0
grpcio==1.32.0
gym==0.18.3
gym-unity==0.26.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.3.1
ipython==7.24.1
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keras-tuner @ git+https://github.com/keras-team/keras-tuner.git@86cff68ae4c7f35bb6557e75ec6a60e0766454d2
keras2onnx==1.7.0
kiwisolver==1.3.1
locket==0.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mixpanel==4.7.0
mlagents-envs==0.26.0
msrest==0.6.21
networkx==2.5.1
numpy==1.19.2
oauthlib==3.1.0
onnx==1.6.0
onnxconverter-common==1.8.1
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
parso==0.8.2
partd==1.2.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.18
protobuf==3.17.1
psutil==5.8.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
PyGithub==1.55
pyglet==1.5.15
Pygments==2.9.0
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pywin32==301
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.4.2
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
sentry-sdk==0.14.3
six==1.15.0
smmap==4.0.0
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-addons==0.13.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
tf2onnx==1.7.2
threadpoolctl==2.1.0
tifffile==2021.4.8
toolz==0.11.1
traitlets==5.0.5
typeguard==2.12.1
typing-extensions==3.7.4.3
uamqp==1.4.0
urllib3==1.25.10
wcwidth==0.2.5
websockets==9.1
Werkzeug==2.0.1
wrapt==1.12.1
xgboost==1.4.2
zipp==3.4.1
